// src/Types/types.tsx
export type DataType={
    id: number
    date:string
    firstNameLastName: string
    phone: string
    email: string
    startDate: string
    meals: number | number
    invoice: boolean
}


export type FilterDataType={
        ID: number
        Date:string
        FirstNameLastName: string
        Phone: string
        Email: string
        StartDate: string
        Meals: number | number
        Invoice: boolean
}